refactor/mods: organize tests more idiomatically
