+++ src/data/map.rs
@@ -7,11 +7,11 @@
 // specific language governing permissions and limitations relating to use of the SAFE Network
 // Software.
 
-use crate::authorization::access_control::{
+use crate::data::access_control::{
     AccessListTrait, AccessType, PrivateAccessList, PrivateUserAccess, PublicAccessList,
     PublicUserAccess,
 };
-use crate::shared_data::{
+use crate::shared_types::{
     to_absolute_range, to_absolute_version, Address, ExpectedVersions, Guarded, Key, KeyValuePair,
     Keys, Kind, NonGuarded, Owner, User, Value, Values, Version, CURRENT_VERSION,
 };
